+++ src/Utilities/ColumnUtilities.php
@@ -3,9 +3,14 @@
 namespace Rappasoft\LaravelLivewireTables\Utilities;
 
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
-use Illuminate\Database\Query\Builder as Builder;
+use Illuminate\Database\Query\Builder;
 use Illuminate\Support\Str;
 
+/**
+ * Class ColumnUtilities
+ *
+ * @package Rappasoft\LaravelLivewireTables\Utilities
+ */
 class ColumnUtilities
 {
     /**
@@ -14,7 +19,7 @@ class ColumnUtilities
      * @param $column
      * @return bool
      */
-    public static function hasRelation($column)
+    public static function hasRelation($column): bool
     {
         return Str::contains($column, '.');
     }
@@ -25,7 +30,7 @@ class ColumnUtilities
      * @param $column
      * @return string
      */
-    public static function parseRelation($column)
+    public static function parseRelation($column): string
     {
         return Str::beforeLast($column, '.');
     }
@@ -36,7 +41,7 @@ class ColumnUtilities
      * @param $column
      * @return string
      */
-    public static function parseField($column)
+    public static function parseField($column): string
     {
         return Str::afterLast($column, '.');
     }
@@ -48,9 +53,9 @@ class ColumnUtilities
      * @param $searchColumns
      * @return bool
      */
-    public static function hasMatch($column, $searchColumns)
+    public static function hasMatch($column, $searchColumns): bool
     {
-        return array_search($column, $searchColumns ?? []) !== false;
+        return in_array($column, $searchColumns ?? [], true);
     }
 
     /**
@@ -60,42 +65,44 @@ class ColumnUtilities
      * @param $searchColumns
      * @return bool
      */
-    public static function hasWildcardMatch($column, $searchColumns)
+    public static function hasWildcardMatch($column, $searchColumns): bool
     {
         return count(array_filter($searchColumns ?? [], function ($searchColumn) use ($column) {
-
-                // match wildcards such as * or table.*
+            // Match wildcards such as * or table.*
             $hasWildcard = Str::endsWith($searchColumn, '*');
 
-            // if no wildcard, skip
+            // If no wildcard, skip
             if (! $hasWildcard) {
                 return false;
             }
 
             if (! self::hasRelation($column)) {
                 return true;
-            } else {
-                $selectColumnPrefix = self::parseRelation($searchColumn);
-                $columnPrefix = self::parseRelation($column);
-
-                return $selectColumnPrefix === $columnPrefix;
             }
+
+            $selectColumnPrefix = self::parseRelation($searchColumn);
+            $columnPrefix = self::parseRelation($column);
+
+            return $selectColumnPrefix === $columnPrefix;
         })) > 0;
     }
 
     /**
-     * @param EloquentBuilder|Builder $queryBuilder
-     * @return null
+     * @param  null  $queryBuilder
+     *
+     * @return array|null
      */
-    public static function columnsFromBuilder($queryBuilder = null)
+    public static function columnsFromBuilder($queryBuilder = null): ?array
     {
         if ($queryBuilder instanceof EloquentBuilder) {
             return $queryBuilder->getQuery()->columns;
-        } elseif ($queryBuilder instanceof Builder) {
-            return $queryBuilder->columns;
-        } else {
-            return null;
         }
+
+        if ($queryBuilder instanceof Builder) {
+            return $queryBuilder->columns;
+        }
+
+        return null;
     }
 
     /**
@@ -105,17 +112,17 @@ class ColumnUtilities
      * @param $queryBuilder
      * @return string
      */
-    public static function mapToSelected($column, $queryBuilder)
+    public static function mapToSelected($column, $queryBuilder): ?string
     {
-        // grab select
+        // Grab select
         $select = self::columnsFromBuilder($queryBuilder);
 
-        // can't match anything if no select
+        // Can't match anything if no select
         if (is_null($select)) {
             return null;
         }
 
-        // search builder select for a match
+        // Search builder select for a match
         $hasMatch = self::hasMatch($column, $select);
 
         // example 2 - match
@@ -123,12 +130,12 @@ class ColumnUtilities
         // select: service_statuses.name
         // maps to: service_statuses.name
 
-        // if we found a match, lets use that instead of searching relations
+        // If we found a match, lets use that instead of searching relations
         if ($hasMatch) {
             return $column;
         }
 
-        // search builder select for a wildcard match
+        // Search builder select for a wildcard match
         $hasWildcardMatch = self::hasWildcardMatch($column, $select);
 
         // example 3 - wildcard match
@@ -136,17 +143,17 @@ class ColumnUtilities
         // select: service_statuses.*
         // maps to: service_statuses.name
 
-        // if we found a wildcard match, lets use that instead of matching relations
+        // If we found a wildcard match, lets use that instead of matching relations
         if ($hasWildcardMatch) {
             return $column;
         }
 
-        // split the relation and field
+        // Split the relation and field
         $hasRelation = self::hasRelation($column);
         $relationName = self::parseRelation($column);
         $fieldName = self::parseField($column);
 
-        // we know there is a relation and we know it doesn't match any of the
+        // We know there is a relation and we know it doesn't match any of the
         // select columns. Let's try to grab the table name for the relation
         // and see if that matches something in the select
         //
@@ -155,49 +162,45 @@ class ColumnUtilities
         // select: service_statuses.name
         // maps to: service_statuses.name
 
-        // if we didn't previously match the column and there isn't a relation
+        // If we didn't previously match the column and there isn't a relation
         if (! $hasRelation) {
-
-            // there's nothing else to do
+            // There's nothing else to do
             return null;
+        }
 
-        // this is easiest when using the eloquent query builder
-        } elseif ($queryBuilder instanceof EloquentBuilder) {
+        // This is easiest when using the eloquent query builder
+        if ($queryBuilder instanceof EloquentBuilder) {
             $relation = $queryBuilder->getRelation($relationName);
             $possibleTable = $relation->getModel()->getTable();
         } elseif ($queryBuilder instanceof Builder) {
-
-            // @todo: possible ways to do this?
+            // TODO: Possible ways to do this?
             $possibleTable = null;
         } else {
-
-            // we would have already returned before this is possible
+            // We would have already returned before this is possible
             $possibleTable = null;
         }
 
-        // if we found a possible table
+        // If we found a possible table
         if (! is_null($possibleTable)) {
-
-            // build possible selected column
+            // Build possible selected column
             $possibleSelectColumn = $possibleTable . '.' . $fieldName;
 
             $possibleMatch = self::hasMatch($possibleSelectColumn, $select);
 
-            // if we found a possible match for a relation to an already selected
-            // column, let's use that
+            // If we found a possible match for a relation to an already selected column, let's use that
             if ($possibleMatch) {
                 return $possibleSelectColumn;
             }
 
             $possibleWildcardMatch = self::hasWildcardMatch($possibleSelectColumn, $select);
 
-            // ditto with a possible wildcard match
+            // Ditto with a possible wildcard match
             if ($possibleWildcardMatch) {
                 return $possibleSelectColumn;
             }
         }
 
-        // we couldn't match to a selected column
+        // We couldn't match to a selected column
         return null;
     }
 }